requirements.txt
Flask==3.0.3
pillow==10.3.0
pdf2image==1.17.0
pytesseract==0.3.10
deep-translator==1.11.4

numpy==1.26.4
pandas==2.1.4
scikit-learn==1.3.2

# spaCy stable version with no BLIS/THINC errors on Windows
spacy==3.7.2
blis==0.7.10
thinc==8.2.1
cymem==2.0.8
preshed==3.0.9
srsly==2.4.8
catalogue==2.0.10
